Split code in modules, finish text input
