+++ pygame_vkeyboard/examples/textinput.py
@@ -23,7 +23,7 @@ def main(test=False):
     # Init pygame
     pygame.init()
     screen = pygame.display.set_mode((300, 400))
-    screen.fill((200, 200, 200))
+    screen.fill((20, 100, 100))
 
     # Create keyboard
     layout = vkboard.VKeyboardLayout(vkboard.VKeyboardLayout.QWERTY)
@@ -44,7 +44,6 @@ def main(test=False):
                 exit()
 
         keyboard.update(events)
-
         keyboard.draw()
 
         # Flip surface
